Refactor auth + encryption
Fix a bug with client account token caching
