+++ src/auth/encryption.js
@@ -1,28 +1,25 @@
+const { Ber } = require('asn1')
 const JWT = require('jsonwebtoken')
 const crypto = require('crypto')
-const { Ber } = require('asn1')
 const ecPem = require('ec-pem')
-const fs = require('fs')
-const DataProvider = require('../../data/provider')
+const debug = require('debug')('minecraft-protocol')
 
 const SALT = '🧂'
 const curve = 'secp384r1'
 
 function Encrypt (client, server, options) {
-  const skinGeom = fs.readFileSync(DataProvider(options.protocolVersion).getPath('skin_geom.txt'), 'utf-8')
-
   client.ecdhKeyPair = crypto.createECDH(curve)
   client.ecdhKeyPair.generateKeys()
   client.clientX509 = writeX509PublicKey(client.ecdhKeyPair.getPublicKey())
 
   function startClientboundEncryption (publicKey) {
-    console.warn('[encrypt] Pub key base64: ', publicKey)
+    debug('[encrypt] Client pub key base64: ', publicKey)
     const pubKeyBuf = readX509PublicKey(publicKey.key)
 
     const alice = client.ecdhKeyPair
     const alicePEM = ecPem(alice, curve) // https://github.com/nodejs/node/issues/15116#issuecomment-384790125
     const alicePEMPrivate = alicePEM.encodePrivateKey()
-    // Shared secret from bob's public key + our private key
+    // Shared secret from the client's public key + our private key
     client.sharedSecret = alice.computeSecret(pubKeyBuf)
 
     // Secret hash we use for packet encryption:
@@ -34,10 +31,10 @@ function Encrypt (client, server, options) {
     const secretHash = crypto.createHash('sha256')
     secretHash.update(SALT)
     secretHash.update(client.sharedSecret)
-    console.log('[encrypt] Shared secret', client.sharedSecret)
+    // console.log('[encrypt] Shared secret', client.sharedSecret)
 
     client.secretKeyBytes = secretHash.digest()
-    console.log('[encrypt] Shared hash', client.secretKeyBytes)
+    // console.log('[encrypt] Shared hash', client.secretKeyBytes)
     const x509 = writeX509PublicKey(alice.getPublicKey())
     const token = JWT.sign({
       salt: toBase64(SALT),
@@ -56,7 +53,7 @@ function Encrypt (client, server, options) {
   }
 
   function startServerboundEncryption (token) {
-    console.warn('[encrypt] Starting serverbound encryption', token)
+    debug('[encrypt] Starting serverbound encryption', token)
     const jwt = token?.token
     if (!jwt) {
       // TODO: allow connecting to servers without encryption
@@ -69,7 +66,7 @@ function Encrypt (client, server, options) {
     const body = JSON.parse(String(payload))
     const serverPublicKey = readX509PublicKey(head.x5u)
     client.sharedSecret = alice.computeSecret(serverPublicKey)
-    console.log('[encrypt] Shared secret', client.sharedSecret)
+    // console.log('[encrypt] Shared secret', client.sharedSecret)
 
     const salt = Buffer.from(body.salt, 'base64')
 
@@ -78,7 +75,7 @@ function Encrypt (client, server, options) {
     secretHash.update(client.sharedSecret)
 
     client.secretKeyBytes = secretHash.digest()
-    console.log('[encrypt] Shared hash', client.secretKeyBytes)
+    // console.log('[encrypt] Shared hash', client.secretKeyBytes)
     const initial = client.secretKeyBytes.slice(0, 16)
     client.startEncryption(initial)
 
@@ -89,70 +86,6 @@ function Encrypt (client, server, options) {
 
   client.on('server.client_handshake', startClientboundEncryption)
   client.on('client.server_handshake', startServerboundEncryption)
-
-  client.createClientChain = (mojangKey) => {
-    mojangKey = mojangKey || require('./constants').PUBLIC_KEY
-    const alice = client.ecdhKeyPair
-    const alicePEM = ecPem(alice, curve) // https://github.com/nodejs/node/issues/15116#issuecomment-384790125
-    const alicePEMPrivate = alicePEM.encodePrivateKey()
-
-    const token = JWT.sign({
-      identityPublicKey: mojangKey,
-      certificateAuthority: true
-    }, alicePEMPrivate, { algorithm: 'ES384', header: { x5u: client.clientX509 } })
-
-    client.clientIdentityChain = token
-    client.createClientUserChain(alicePEMPrivate)
-  }
-
-  client.createClientUserChain = (privateKey) => {
-    let payload = {
-      ServerAddress: options.hostname,
-      ThirdPartyName: client.profile.name,
-      DeviceOS: client.session?.deviceOS || 1,
-      GameVersion: options.version || '1.16.201',
-      ClientRandomId: Date.now(), // TODO make biggeer
-      DeviceId: '2099de18-429a-465a-a49b-fc4710a17bb3', // TODO random
-      LanguageCode: 'en_GB', // TODO locale
-      AnimatedImageData: [],
-      PersonaPieces: [],
-      PieceTintColours: [],
-      SelfSignedId: '78eb38a6-950e-3ab9-b2cf-dd849e343701',
-      SkinId: '5eb65f73-af11-448e-82aa-1b7b165316ad.persona-e199672a8c1a87e0-0',
-      SkinData: 'AAAAAA==',
-      SkinResourcePatch: 'ewogICAiZ2VvbWV0cnkiIDogewogICAgICAiYW5pbWF0ZWRfMTI4eDEyOCIgOiAiZ2VvbWV0cnkuYW5pbWF0ZWRfMTI4eDEyOF9wZXJzb25hLWUxOTk2NzJhOGMxYTg3ZTAtMCIsCiAgICAgICJhbmltYXRlZF9mYWNlIiA6ICJnZW9tZXRyeS5hbmltYXRlZF9mYWNlX3BlcnNvbmEtZTE5OTY3MmE4YzFhODdlMC0wIiwKICAgICAgImRlZmF1bHQiIDogImdlb21ldHJ5LnBlcnNvbmFfZTE5OTY3MmE4YzFhODdlMC0wIgogICB9Cn0K',
-      SkinGeometryData: skinGeom,
-      SkinImageHeight: 1,
-      SkinImageWidth: 1,
-      ArmSize: 'wide',
-      CapeData: '',
-      CapeId: '',
-      CapeImageHeight: 0,
-      CapeImageWidth: 0,
-      CapeOnClassicSkin: false,
-      PlatformOfflineId: '',
-      PlatformOnlineId: '', // chat
-      // a bunch of meaningless junk
-      CurrentInputMode: 1,
-      DefaultInputMode: 1,
-      DeviceModel: '',
-      GuiScale: -1,
-      UIProfile: 0,
-      TenantId: '',
-      PremiumSkin: false,
-      PersonaSkin: false,
-      PieceTintColors: [],
-      SkinAnimationData: '',
-      ThirdPartyNameOnly: false,
-      SkinColor: '#ffffcd96'
-    }
-    payload = require('./logPack.json')
-    const customPayload = options.userData || {}
-    payload = { ...payload, ...customPayload }
-
-    client.clientUserChain = JWT.sign(payload, privateKey,
-      { algorithm: 'ES384', header: { x5u: client.clientX509 } })
-  }
 }
 
 function toBase64 (string) {
