+++ src/client/tokens.js
@@ -13,12 +13,7 @@ class MsaTokenManager {
     this.scopes = scopes
     this.cacheLocation = cacheLocation || path.join(__dirname, './msa-cache.json')
 
-    try {
-      this.msaCache = require(this.cacheLocation)
-    } catch (e) {
-      this.msaCache = {}
-      fs.writeFileSync(this.cacheLocation, JSON.stringify(this.msaCache))
-    }
+    this.reloadCache()
 
     const beforeCacheAccess = async (cacheContext) => {
       cacheContext.tokenCache.deserialize(await fs.promises.readFile(this.cacheLocation, 'utf-8'))
@@ -40,6 +35,15 @@ class MsaTokenManager {
     }
     this.msalApp = new msal.PublicClientApplication(msalConfig)
     this.msalConfig = msalConfig
+  }
+
+  reloadCache () {
+    try {
+      this.msaCache = require(this.cacheLocation)
+    } catch (e) {
+      this.msaCache = {}
+      fs.writeFileSync(this.cacheLocation, JSON.stringify(this.msaCache))
+    }
   }
 
   getUsers () {
@@ -89,6 +93,7 @@ class MsaTokenManager {
     return new Promise((resolve, reject) => {
       this.msalApp.acquireTokenByRefreshToken(refreshTokenRequest).then((response) => {
         debug('[msa] refreshed token', JSON.stringify(response))
+        this.reloadCache()
         resolve(response)
       }).catch((error) => {
         debug('[msa] failed to refresh', JSON.stringify(error))
